==> 00_tweet_download.ipynb <==
## Link utili:

- Homepage `twint`: https://github.com/twintproject/twint

- Installazione della versione corretta della libreria `tornado`: https://stackoverflow.com/questions/53248431/asyncio-runtimeerror-this-event-loop-is-already-running

- Campi di ricerca: https://github.com/twintproject/twint/wiki/Configuration

- Formattazione: https://github.com/twintproject/twint/wiki/Tweet-attributes

- `twint` e `Pandas`: https://github.com/twintproject/twint/wiki/Pandas-integration

- Indexing and selecting data in `Pandas`: https://pandas.pydata.org/pandas-docs/stable/user_guide/indexing.html

- Homepage `emoji`: https://pypi.org/project/emoji/

- Lista completa emoji: http://www.unicode.org/emoji/charts/full-emoji-list.html

import twint
import pandas as pd
from datetime import datetime
from emoji import demojize

Lista delle persone escluse:
- [M5S] Beppe Grillo
- [PD] Paola Boldrini (boldrini_paola) ha solo 21 tweet
- [LSP] Lega - Salvini Premier (LegaSalvini) ha troppi tweet
- [LSP] Erika Stefani (erikastefani71) ha solo 10 tweet
- [PD] Walter Verini (VeriniWalter) non posta nulla dal 2018

Persone rimosse per bilanciare il dataset:
- [FdI] "FratellidItalia": "Fratelli d'Italia"
- [M5S] "SPatuanelli": "Stefano Patuanelli"
- [LSP] "MolinariRik": "Riccardo Molinari"
- [LSP] "FabriCecchetti": "Fabrizio Cecchetti"
- [FdI] "Ignazio_LaRussa": "Ignazio La Russa"
- [FdI] "RaffaeleFitto": "Raffaele Fitto"

# Movimento 5 Stelle
M5S = {"Mov5Stelle": "MoVimento 5 Stelle",
       "luigidimaio": "Luigi di Maio",
       "Roberto_Fico": "Roberto Fico",
       "LauraBottici": "Laura Bottici",
       "vitocrimi": "Vito Crimi",
       "virginiaraggi": "Virginia Raggi"}
# Lega - Salvini Premier
LSP = {"matteosalvinimi": "Matteo Salvini",
       "Fontana3Lorenzo": "Lorenzo Fontana",
       "massimogara": "Massimo Garavaglia"}
# Partito Democratico
PD = {"pdnetwork": "Partito Democratico",
      "EnricoLetta": "Enrico Letta",
      "itinagli": "Irene Tinagli",
      "peppeprovenzano": "Giuseppe Provenzano",
      "F_Boccia": "Francesco Boccia",
      "lauraboldrini": "Laura Boldrini",
      "PaoloGentiloni": "Paolo Gentiloni",
      "robersperanza": "Roberto Speranza",
      "dariofrance": "Dario Franceschini",
      "nzingaretti": "Nicola Zingaretti",
      "ZanAlessandro": "Alessandro Zan"}
# Forza Italia
FI = {"forza_italia": "Forza Italia",
      "berlusconi": "Silvio Berlusconi",
      "Antonio_Tajani": "Antonio Tajani",
      "msgelmini": "Mariastella Gelmini",
      "mara_carfagna": "Mara Carfagna",
      "renatobrunetta": "Renato Brunetta",
      "gasparripdl": "Maurizio Gasparri",
      "BerniniAM": "Anna Maria Bernini"}
# Fratelli d'Italia
FdI = {"GiorgiaMeloni": "Giorgia Meloni",
       "GuidoCrosetto": "Guido Crosetto",
       "DSantanche": "Daniela Santanchè"}
# Italia Viva
IV = {"ItaliaViva": "Italia Viva",
      "matteorenzi": "Matteo Renzi",
      "meb": "Maria Elena Boschi",
      "Ettore_Rosato": "Ettore Rosato",
      "TeresaBellanova": "Teresa Bellanova",
      "elenabonetti": "Elena Bonetti",
      "davidefaraone": "Davide Faraone"}

partito_list = {"M5S": list(M5S.keys()),
                "LSP": list(LSP.keys()),
                "PD": list(PD.keys()),
                "FI": list(FI.keys()),
                "FdI": list(FdI.keys()),
                "IV": list(IV.keys())}

Scarico da Twitter tutti i tweet degli utenti contenuti nella lista `username_list`.

start_time = datetime.now()

tweets_df_list = {}

for partito in partito_list:
    username_list = partito_list[partito]
    print(partito + ":", username_list)
    for i in range(len(username_list)):
        user_name = username_list[i]
        print("[" + str(i+1) + "/" + str(len(username_list)) + "]", datetime.now().strftime("%Y-%m-%d %H:%M:%S"))
        try:
            c = twint.Config()

            c.Username = user_name
            c.User_full = True         
            c.Since = "2020-06-03"
            c.Until = "2021-05-15"

            c.Lowercase = True # rende tutto minuscolo

            c.Count = True     # stampo a video il numero di tweet trovati
            c.Pandas = True
            c.Hide_output = True 

            twint.run.Search(c)

            tweets_df_list[user_name] = twint.storage.panda.Tweets_df
            tweets_df_list[user_name]["partito"] = partito
        except Exception as ex:
            print("*"*30)
            print(str(ex))
            print("*"*30)
    print("")

print("" + "*"*30)
print("  Start:", start_time.strftime("%Y-%m-%d %H:%M:%S"))
print("    End:", datetime.now().strftime("%Y-%m-%d %H:%M:%S"))
print("*"*30)

Unisco i vari dataframe contenuti nella lista `tweet_df_list` in uno unico che andrà a formare il *corpus*.

for user_name in tweets_df_list:
    print(user_name.ljust(16), str(tweets_df_list[user_name].shape).rjust(13))
print("-"*30)

# creo un unico dataframe
tweets = pd.concat(tweets_df_list.values())
print("".ljust(16), str(tweets.shape).rjust(13))

Esporto il dataset originale in formato *csv*.

tweets.to_csv("data/tweets_original_"+datetime.now().strftime("%Y_%m_%d")+".csv", sep="\t", index=False, encoding='utf-8-sig', mode="w")

Gestisco gli emoji convertendoli in stringhe di testo comprensibile. Come deliminatore si utilizza `(" emote_"," ")` in modo da rendere le emoji facilmente distinguibili; gli spazi servono per essere sicuri di considerarle come una singola entità.

tweets["tweet"] = tweets["tweet"].apply(lambda x: demojize(x, language="en", delimiters=(" emote_"," ")))

Infine, esporto il dataset normalizzato in formato *csv*.

tweets.to_csv("data/00_tweets.csv", sep="\t", index=False, encoding='utf-8-sig', mode="w")
